llvm: Disable the stock llvm if QCOM's optimized LLVM is used[JB_2.5]
Customers can choose the stock LLVM or QCOM's optimized LLVM in
BoardConfig.mk. When BOARD_USE_QCOM_LLVM_CLANG_RS is set to be
"true", the stock LLVM will be disabled and QCOM's LLVM will be
enabled. The functions in llvm-select.mk returns the
corresponding path of the selected LLVM based on customer's selection.

Change-Id: Ia70e881fc9d7733763f2267f0f9aa5e0b1924741